refactor: entity table
