+++ packages/web-app/src/components/appli/AdvancedSearch/OrganizationsSearch.jsx
@@ -19,7 +19,7 @@ import {
 } from './SearchElements';
 import SearchInput from '../../common/SearchInput';
 import { ADVANCED_SEARCH_TYPES } from '../../../conf/config';
-import { getStoredRowsPerPage } from '../../common/EntityTable/EntityTable';
+import { getStoredRowsPerPage } from '../../common/EntityTable';
 
 const FILTER_LABELS = {
   city: 'City',
