+++ packages/web-app/src/components/appli/AdvancedSearch/PersonSearch.jsx
@@ -9,7 +9,7 @@ import {
 import { SearchForm, SearchActionButtons } from './SearchElements';
 import SearchInput from '../../common/SearchInput';
 import { ADVANCED_SEARCH_TYPES } from '../../../conf/config';
-import { getStoredRowsPerPage } from '../../common/EntityTable/EntityTable';
+import { getStoredRowsPerPage } from '../../common/EntityTable';
 
 const PersonSearch = () => {
   const dispatch = useDispatch();
