+++ packages/web-app/src/components/appli/Duplicates/DuplicatesList.jsx
@@ -5,7 +5,7 @@ import {
   deleteDuplicates,
   fetchDuplicatesList
 } from '../../../actions/DuplicatesImport';
-import EntityTable from '../../common/EntityTable/EntityTable';
+import EntityTable from '../../common/EntityTable';
 
 import TableActions from './TableActions';
 import { resetDocumentApiErrors } from '../../../actions/Document/ResetApiErrors';
