+++ packages/web-app/src/components/common/EntityTable/index.js
@@ -1,0 +1,1 @@
+export { default, getStoredRowsPerPage } from './EntityTable';
